get frame number
